backend/pyproject.toml
[project]
name = "paper-trader"
version = "1.0.0"
description = "Multi-market paper trading platform"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2,<6",
    "djangorestframework>=3.15,<4",
    "djangorestframework-simplejwt>=5.3,<6",
    "django-filter>=24,<25",
    "django-celery-beat>=2.6,<3",
    "channels>=4.2,<5",
    "channels-redis>=4.2,<5",
    "psycopg2-binary>=2.9,<3",
    "redis>=5.0,<6",
    "celery>=5.4,<6",
    "yfinance>=0.2,<1",
    "exchange-calendars>=4.6,<5",
    "pytz>=2024.1",
    "python-decouple>=3.8,<4",
    "gunicorn>=22,<23",
    "uvicorn>=0.30,<1",
    "daphne>=4.1,<5",
    "sentry-sdk>=2.0,<3",
    "django-cors-headers>=4.4,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "factory-boy>=3.3",
    "ruff>=0.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "tests.py test_*.py *_tests.py"